Place DateTimePicker composable in 'composables' package so that it can be reused if needed
